requirements.txt
sentence-transformers
llama-cpp-python
PyMuPDF
torch
huggingface-hub
